Correct semantically atm branches and subsidary sections.
